Advanced MFW breakdown part 8/8
Still doesn't work on the oad process of the TBM, I can't figure out why the solver oscillate between two values
