+++ src/fastga/models/load_analysis/wing/aerostructural_loads.py
@@ -15,6 +15,8 @@ Computes the aerostructural loads on the wing of the aircraft.
 #  You should have received a copy of the GNU General Public License
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+import warnings
+
 import numpy as np
 import openmdao.api as om
 from scipy.integrate import trapz
@@ -24,8 +26,8 @@ from stdatm import Atmosphere
 import fastoad.api as oad
 
 from fastga.models.aerodynamics.constants import SPAN_MESH_POINT
-from fastga.models.geometry.geom_components.wing_tank.compute_mfw_advanced import (
-    tank_volume_distribution,
+from fastga.models.geometry.geom_components.wing_tank.wing_tank_components.compute_wing_tank_y_array import (
+    POINTS_NB_WING,
 )
 
 from .constants import SUBMODEL_AEROSTRUCTURAL_LOADS, NB_POINTS_POINT_MASS, POINT_MASS_SPAN_RATIO
@@ -692,3 +694,90 @@ class AerostructuralLoad(om.ExplicitComponent):
         chord_vector_new = chord_vector
 
         return y_vector_new, chord_vector_new, point_mass_array_new
+
+
+def tank_volume_distribution(inputs, y_array_orig):
+
+    root_chord = inputs["data:geometry:wing:root:chord"]
+    tip_chord = inputs["data:geometry:wing:tip:chord"]
+    root_y = inputs["data:geometry:wing:root:y"]
+    tip_y = inputs["data:geometry:wing:tip:y"]
+    root_tc = inputs["data:geometry:wing:root:thickness_ratio"]
+    tip_tc = inputs["data:geometry:wing:tip:thickness_ratio"]
+    flap_chord_ratio = inputs["data:geometry:flap:chord_ratio"]
+    aileron_chord_ratio = inputs["data:geometry:wing:aileron:chord_ratio"]
+    y_ratio_tank_beginning = inputs["data:geometry:propulsion:tank:y_ratio_tank_beginning"]
+    y_ratio_tank_end = inputs["data:geometry:propulsion:tank:y_ratio_tank_end"]
+    engine_config = inputs["data:geometry:propulsion:engine:layout"]
+    le_chord_percentage = inputs["data:geometry:propulsion:tank:LE_chord_percentage"]
+    te_chord_percentage = inputs["data:geometry:propulsion:tank:TE_chord_percentage"]
+    nacelle_width = inputs["data:geometry:propulsion:nacelle:width"]
+    span = inputs["data:geometry:wing:span"]
+    lg_type = inputs["data:geometry:landing_gear:type"]
+    y_lg = inputs["data:geometry:landing_gear:y"]
+    k = inputs["settings:geometry:fuel_tanks:depth"]
+
+    # Create the array which will contain the tank cross section area at each section
+
+    y_array = y_array_orig.flatten()
+
+    semi_span = span / 2
+    y_tank_beginning = semi_span * y_ratio_tank_beginning
+    y_tank_end = semi_span * y_ratio_tank_end
+
+    y_in_tank_index = np.where((y_array >= y_tank_beginning) & (y_array <= y_tank_end))[0]
+    y_in_tank_array = y_array[y_in_tank_index]
+
+    slope_chord = (tip_chord - root_chord) / (tip_y - root_y)
+
+    chord_array = np.zeros_like(y_array)
+    for y in y_in_tank_array:
+        idx = np.where(y_array == y)[0]
+        if y < root_y:
+            chord_array[idx] = root_chord
+        else:
+            chord_array[idx] = slope_chord * y + root_chord
+
+    # Computation of the thickness ratio profile along the span, as tc = slope * y +
+    # tc_fuselage_center.
+    slope_tc = (tip_tc - root_tc) / (tip_y - root_y)
+    thickness_ratio_array = np.zeros_like(y_array)
+    for y in y_in_tank_array:
+        idx = np.where(y_array == y)[0]
+        if y < root_y:
+            thickness_ratio_array[idx] = root_tc
+        else:
+            thickness_ratio_array[idx] = slope_tc * y + root_tc
+
+    # The k factor stating the depth of the fuel tanks is included here.
+    thickness_array = k * chord_array * thickness_ratio_array
+
+    # Distributed propulsion / single engine on the wing / nose or rear mounted engine
+    if engine_config != 1.0:
+        y_ratio = 0.0
+    else:
+        y_ratio = inputs["data:geometry:propulsion:engine:y_ratio"]
+
+    y_eng_array = semi_span * np.array(y_ratio)
+
+    in_eng_nacelle = np.full(len(y_in_tank_array), False)
+    for y_eng in y_eng_array:
+        for i in np.where(abs(y_in_tank_array - y_eng) <= nacelle_width / 2.0):
+            in_eng_nacelle[i] = True
+    where_engine = np.where(in_eng_nacelle)
+
+    width_array = (
+        1.0 - le_chord_percentage - te_chord_percentage - max(flap_chord_ratio, aileron_chord_ratio)
+    ) * chord_array
+    if engine_config == 1.0:
+        for i in where_engine:
+            # For now 50% size reduction in the fuel tank capacity due to the engine
+            width_array[i] *= 0.5
+    if lg_type == 1.0:
+        for i in np.where(y_array < y_lg):
+            # For now 80% size reduction in the fuel tank capacity due to the landing gear
+            width_array[i] *= 0.2
+
+    tank_cross_section = thickness_array * width_array * 0.85
+
+    return tank_cross_section
